Fixes bug related to Workload type PRESET_IN_FILE
The queue that keeps messge sizes and interarrival times was typed to
keep interarrival times as unit64_t numbers instead of double numbers.
This bug in now fixed.

+++ RpcTransportDesign/OMNeT++Simulation/homatransport/src/application/MsgSizeDistributions.h
@@ -46,7 +46,7 @@ class MsgSizeDistributions {
    
   private:
     std::vector<std::pair<int, double>> msgSizeProbDistVector;
-    std::queue<std::pair<int, uint64_t>> msgSizeInterarrivalQueue;
+    std::queue<std::pair<int, double>> msgSizeInterarrivalQueue;
 
     // Random number generator classes and datastructs
     std::random_device rdDevice;
